chore(mempool_node): move node config
commit-id:39cafc95
